+++ pynfft/nfft.pyx
@@ -19,18 +19,29 @@ import numpy as np
 cimport numpy as np
 from libc.stdlib cimport malloc, free
 from libc cimport limits
-from cnfft3 cimport (nfft_adjoint, nfft_adjoint_direct, nfft_init_guru,
-                     nfft_trafo, nfft_trafo_direct, nfft_precompute_one_psi,
-                     nfft_finalize, fftw_complex, nfft_plan)
-from cnfft3 cimport (PRE_PHI_HUT, FG_PSI, PRE_LIN_PSI, PRE_FG_PSI, PRE_PSI,
-                     PRE_FULL_PSI, MALLOC_X, MALLOC_F_HAT, MALLOC_F,
-                     FFT_OUT_OF_PLACE, FFTW_INIT, NFFT_SORT_NODES,
-                     NFFT_OMP_BLOCKWISE_ADJOINT, PRE_ONE_PSI, FFTW_ESTIMATE,
-                     FFTW_DESTROY_INPUT,)
-from cnfft3 cimport fftw_init_threads, fftw_cleanup_threads
+from cnfft3 cimport *
 
 
-# expose flag management internals for testing
+# Initialize module
+# Numpy must be initialized. When using numpy from C or Cython you must
+# _always_ do that, or you will have segfaults
+np.import_array()
+
+# initialize FFTW threads
+fftw_init_threads()
+
+# register FFTW threads cleanup routine
+import atexit
+@atexit.register
+def _cleanup():
+    fftw_cleanup_threads()
+
+
+########
+# NFFT #
+########
+
+cdef object nfft_supported_flags_tuple
 nfft_supported_flags_tuple = (
     'PRE_PHI_HUT',
     'FG_PSI',
@@ -41,6 +52,7 @@ nfft_supported_flags_tuple = (
     )
 nfft_supported_flags = nfft_supported_flags_tuple
 
+cdef object nfft_flags_dict
 nfft_flags_dict = {
     'PRE_PHI_HUT':PRE_PHI_HUT,
     'FG_PSI':FG_PSI,
@@ -59,170 +71,13 @@ nfft_flags_dict = {
     }
 nfft_flags = nfft_flags_dict.copy()
 
+cdef object fftw_flags_dict
 fftw_flags_dict = {
     'FFTW_ESTIMATE':FFTW_ESTIMATE,
     'FFTW_DESTROY_INPUT':FFTW_DESTROY_INPUT,
     }
 fftw_flags = fftw_flags_dict.copy()
 
-cdef void *nfft_init_double(int d, int *N, int M, int *n, int m,
-                            unsigned nfft_flags, unsigned fftw_flags):
-    cdef nfft_plan *ths = <nfft_plan *>malloc(sizeof(nfft_plan))
-    if ths != NULL:
-        nfft_init_guru(ths, d, N, M, n, m, nfft_flags, fftw_flags)
-    return ths
-
-cdef void nfft_finalize_double(void *_plan):
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_finalize(ths)
-
-cdef void nfft_precompute_double(void *_plan) nogil:
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_precompute_one_psi(ths)
-
-cdef void nfft_trafo_double(void *_plan) nogil:
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_trafo(ths)
-
-cdef void nfft_trafo_direct_double(void *_plan) nogil:
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_trafo_direct(ths)
-
-cdef void nfft_adjoint_double(void *_plan) nogil:
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_adjoint(ths)
-
-cdef void nfft_adjoint_direct_double(void *_plan) nogil:
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    nfft_adjoint_direct(ths)
-
-cdef void nfft_set_x_double(void *_plan, object x):
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    if ths != NULL:
-        ths.x = <double *>np.PyArray_DATA(x)
-
-cdef void nfft_set_f_double(void *_plan, object f):
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    if ths != NULL:
-        ths.f = <fftw_complex *>np.PyArray_DATA(f)
-
-cdef void nfft_set_f_hat_double(void *_plan, object f_hat):
-    cdef nfft_plan *ths = <nfft_plan *> _plan
-    if ths != NULL:
-        ths.f_hat = <fftw_complex *>np.PyArray_DATA(f_hat)
-
-cdef nfft_generic_init nfft_init_per_dtype[1]
-
-cdef nfft_generic_init* _build_nfft_init_list():
-    nfft_init_per_dtype[0] = <nfft_generic_init>&nfft_init_double
-    #nfft_init_per_dtype[1] = <nfft_generic_init>&nfft_init_single
-    #nfft_init_per_dtype[2] = <nfft_generic_init>&nfft_init_ldouble
-
-cdef nfft_generic_finalize nfft_finalize_per_dtype[1]
-
-cdef nfft_generic_finalize* _build_nfft_finalize_list():
-    nfft_finalize_per_dtype[0] = <nfft_generic_finalize>&nfft_finalize_double
-    #nfft_finalize_per_dtype[1] = <nfft_generic_finalize>&nfft_finalize_single
-    #nfft_finalize_per_dtype[2] = <nfft_generic_finalize>&nfft_finalize_ldouble
-
-cdef nfft_generic_precompute nfft_precompute_per_dtype[1]
-
-cdef nfft_generic_precompute* _build_nfft_precompute_list():
-    nfft_precompute_per_dtype[0] = <nfft_generic_precompute>&nfft_precompute_double
-    #nfft_precompute_per_dtype[1] = <nfft_generic_precompute>&nfft_precompute_single
-    #nfft_precompute_per_dtype[2] = <nfft_generic_precompute>&nfft_precompute_ldouble
-
-cdef nfft_generic_trafo nfft_trafo_per_dtype[1]
-
-cdef nfft_generic_trafo* _build_nfft_trafo_list():
-    nfft_trafo_per_dtype[0] = <nfft_generic_trafo>&nfft_trafo_double
-    #nfft_trafo_per_dtype[1] = <nfft_generic_trafo>&nfft_trafo_single
-    #nfft_trafo_per_dtype[2] = <nfft_generic_trafo>&nfft_trafo_ldouble
-
-cdef nfft_generic_trafo_direct nfft_trafo_direct_per_dtype[1]
-
-cdef nfft_generic_trafo_direct* _build_nfft_trafo_direct_list():
-    nfft_trafo_direct_per_dtype[0] = <nfft_generic_trafo_direct>&nfft_trafo_direct_double
-    #nfft_trafo_direct_per_dtype[1] = <nfft_generic_trafo_direct>&nfft_trafo_direct_single
-    #nfft_trafo_direct_per_dtype[2] = <nfft_generic_trafo_direct>&nfft_trafo_direct_ldouble
-
-cdef nfft_generic_adjoint nfft_adjoint_per_dtype[1]
-
-cdef nfft_generic_adjoint* _build_nfft_adjoint_list():
-    nfft_adjoint_per_dtype[0] = <nfft_generic_adjoint>&nfft_adjoint_double
-    #nfft_adjoint_per_dtype[1] = <nfft_generic_adjoint>&nfft_adjoint_single
-    #nfft_adjoint_per_dtype[2] = <nfft_generic_adjoint>&nfft_adjoint_ldouble
-
-cdef nfft_generic_adjoint_direct nfft_adjoint_direct_per_dtype[1]
-
-cdef nfft_generic_adjoint_direct* _build_nfft_adjoint_direct_list():
-    nfft_adjoint_direct_per_dtype[0] = <nfft_generic_adjoint_direct>&nfft_adjoint_direct_double
-    #nfft_adjoint_direct_per_dtype[1] = <nfft_generic_adjoint_direct>&nfft_adjoint_direct_single
-    #nfft_adjoint_direct_per_dtype[2] = <nfft_generic_adjoint_direct>&nfft_adjoint_direct_ldouble
-
-cdef nfft_generic_set_x nfft_set_x_per_dtype[1]
-
-cdef nfft_generic_set_x* _build_nfft_set_x_list():
-    nfft_set_x_per_dtype[0] = <nfft_generic_set_x>&nfft_set_x_double
-    #nfft_set_x_per_dtype[1] = <nfft_generic_set_x>&nfft_set_x_single
-    #nfft_set_x_per_dtype[2] = <nfft_generic_set_x>&nfft_set_x_ldouble
-
-cdef nfft_generic_set_f nfft_set_f_per_dtype[1]
-
-cdef nfft_generic_set_f* _build_nfft_set_f_list():
-    nfft_set_f_per_dtype[0] = <nfft_generic_set_f>&nfft_set_f_double
-    #nfft_set_f_per_dtype[1] = <nfft_generic_set_f>&nfft_set_f_single
-    #nfft_set_f_per_dtype[2] = <nfft_generic_set_f>&nfft_set_f_ldouble
-
-cdef nfft_generic_set_f_hat nfft_set_f_hat_per_dtype[1]
-
-cdef nfft_generic_set_f_hat* _build_nfft_set_f_hat_list():
-    nfft_set_f_hat_per_dtype[0] = <nfft_generic_set_f_hat>&nfft_set_f_hat_double
-    #nfft_set_f_hat_per_dtype[1] = <nfft_generic_set_f_hat>&nfft_set_f_hat_single
-    #nfft_set_f_hat_per_dtype[2] = <nfft_generic_set_f_hat>&nfft_set_f_hat_ldouble
-
-cdef object nfft_complex_dtypes
-nfft_complex_dtypes = {
-        np.dtype('float64'): np.dtype('complex128')
-        #np.dtype('float32'): np.dtype('complex64')
-        #np.dtype('float128'): np.dtype('complex256')
-        }
-
-cdef object nfft_dtype_to_index
-nfft_dtype_to_index = {
-        np.dtype('float64'): 0
-        #np.dtype('float32'): 1
-        #np.dtype('float128'): 2
-        }
-
-# Numpy must be initialized. When using numpy from C or Cython you must
-# _always_ do that, or you will have segfaults
-np.import_array()
-
-# initialize module
-_build_nfft_init_list()
-_build_nfft_finalize_list()
-_build_nfft_precompute_list()
-_build_nfft_trafo_list()
-_build_nfft_trafo_direct_list()
-_build_nfft_adjoint_list()
-_build_nfft_adjoint_direct_list()
-_build_nfft_set_x_list()
-_build_nfft_set_f_list()
-_build_nfft_set_f_hat_list()
-
-# initialize FFTW threads
-fftw_init_threads()
-#fftwf_init_threads()
-#fftwl_init_threads()
-
-# set threads' cleanup routine
-import atexit
-@atexit.register
-def _cleanup():
-    fftw_cleanup_threads()
-#    fftwf_cleanup_threads()
-#    fftwl_cleanup_threads()
 
 cdef class NFFT:
     '''
@@ -239,35 +94,37 @@ cdef class NFFT:
     a ``ValueError`` exception.
 
     The nodes must be initialized prior to precomputing the operator with the
-    :meth:`~pynfft.nfft.NFFT.precompute` method.
+    :meth:`pynfft.NFFT.precompute` method.
 
     The forward and adjoint NFFT operation may be performed by calling the
-    :meth:`~pynfft.nfft.NFFT.trafo` or :meth:`~pynfft.nfft.NFFT.adjoint`
+    :meth:`pynfft.NFFT.trafo` or :meth:`pynfft.NFFT.adjoint`
     methods. The NDFT may also be computed by calling the
-    :meth:`~pynfft.nfft.NFFT.trafo_direct` or
-    :meth:`~pynfft.nfft.NFFT.adjoint_direct`.
+    :meth:`pynfft.NFFT.trafo_direct` or
+    :meth:`pynfft.NFFT.adjoint_direct`.
     '''
+    cdef nfft_plan _plan
+    cdef int _d
+    cdef int _m
+    cdef int _M_total
+    cdef int _N_total
+    cdef object _f
+    cdef object _f_hat
+    cdef object _x
+    cdef object _N
+    cdef object _n
+    cdef object _dtype
+    cdef object _flags
+
     # where the C-related content of the class is being initialized
     def __cinit__(self, N, M, n=None, m=12, x=None, f=None, f_hat=None,
-                  dtype=None, flags=None, *args, **kwargs):
+                  flags=None, *args, **kwargs):
 
-        # check dtype and assign function pointers accordingly
-        dtype = np.dtype('float64') if dtype is None else np.dtype(dtype)
-        try:
-            dtype_complex = nfft_complex_dtypes[dtype]
-            func_idx = nfft_dtype_to_index[dtype]
-            self.__nfft_init = nfft_init_per_dtype[func_idx]
-            self.__nfft_finalize = nfft_finalize_per_dtype[func_idx]
-            self.__nfft_precompute = nfft_precompute_per_dtype[func_idx]
-            self.__nfft_trafo = nfft_trafo_per_dtype[func_idx]
-            self.__nfft_trafo_direct = nfft_trafo_direct_per_dtype[func_idx]
-            self.__nfft_adjoint = nfft_adjoint_per_dtype[func_idx]
-            self.__nfft_adjoint_direct = nfft_adjoint_direct_per_dtype[func_idx]
-            self.__nfft_set_x = nfft_set_x_per_dtype[func_idx]
-            self.__nfft_set_f = nfft_set_f_per_dtype[func_idx]
-            self.__nfft_set_f_hat = nfft_set_f_hat_per_dtype[func_idx]
-        except KeyError:
-            raise ValueError('dtype %s is not supported' % dtype)
+        # support only double / double complex NFFT
+        # TODO: if support for multiple floating precision lands in the
+        # NFFT library, adapt this section to dynamically figure the
+        # real and complex dtypes
+        dtype_real = np.dtype('float64')
+        dtype_complex = np.dtype('complex128')
 
         # NOTE: use of reshape([-1, 1]) to avoid working with 0-d arrays which
         # cannot be indexed explictly
@@ -311,18 +168,18 @@ cdef class NFFT:
         if x is not None:
             if not x.flags.c_contiguous:
                 raise ValueError('x array must be contiguous')
-            if x.dtype != np.float64:
+            if x.dtype != dtype_real:
                 raise ValueError('x must be of type float64')
             if x.size != M_total * d:
                 raise ValueError('x must be of size %d'%(M_total * d))
             self._x = x
         else:
-            self._x = np.empty(M_total*d, dtype=dtype)
+            self._x = np.empty(M_total*d, dtype=dtype_real)
 
         if f is not None:
             if not f.flags.c_contiguous:
                 raise ValueError('f array must be contiguous')
-            if f.dtype != np.complex128:
+            if f.dtype != dtype_complex:
                 raise ValueError('f must be of type float64')
             if f.size != M_total:
                 raise ValueError('f must be of size %d'%(M_total))
@@ -333,7 +190,7 @@ cdef class NFFT:
         if f_hat is not None:
             if not f_hat.flags.c_contiguous:
                 raise ValueError('f_hat array must be contiguous')
-            if f_hat.dtype != np.complex128:
+            if f_hat.dtype != dtype_complex:
                 raise ValueError('f_hat must be of type float64')
             if f_hat.size != N_total:
                 raise ValueError('f_hat must be of size %d'%(N_total))
@@ -407,32 +264,32 @@ cdef class NFFT:
             _n[t] = n[t, 0]
 
         try:
-            self. __plan = self.__nfft_init(_d, _N, _M_total, _n, _m,
+            nfft_init_guru(&self._plan, _d, _N, _M_total, _n, _m,
                     _nfft_flags, _fftw_flags)
-            # in case malloc failed
-            if self.__plan == NULL:
-                raise MemoryError
         except:
-            raise
+            raise MemoryError
         finally:
             free(_N)
             free(_n)
 
-        self.__nfft_set_x(self.__plan, self._x)
-        self.__nfft_set_f(self.__plan, self._f)
-        self.__nfft_set_f_hat(self.__plan, self._f_hat)
+        self._plan.x = (
+            <double *>np.PyArray_DATA(self._x))
+        self._plan.f = (
+            <fftw_complex *>np.PyArray_DATA(self._f))
+        self._plan.f_hat = (
+            <fftw_complex *>np.PyArray_DATA(self._f_hat))
         self._d = d
         self._m = m[0, 0]
         self._M_total = M_total
         self._N_total = N_total
         self._N = tuple([N[t, 0] for t in range(d)])
         self._n = tuple([n[t, 0] for t in range(d)])
-        self._dtype = dtype
+        self._dtype = dtype_real
         self._flags = flags_used
 
     # here, just holds the documentation of the class constructor
     def __init__(self, N, M, n=None, m=12, x=None, f=None, f_hat=None,
-                 dtype=None, flags=None, *args, **kwargs):
+                 flags=None, *args, **kwargs):
         '''
         :param N: multi-bandwith size.
         :type N: int, tuple of int
@@ -448,18 +305,8 @@ cdef class NFFT:
         :type f: ndarray
         :param f_hat: external array holding the Fourier coefficients.
         :type f_hat: ndarray
-        :param dtype: floating precision, see note below.
-        :type dtype: str, numpy.dtype
         :param flags: list of precomputation flags, see note below.
         :type flags: tuple
-
-        **Floating precision**
-
-        Parameter ``dtype`` allows to specify the desired floating point
-        precision. It defaults to None and should not be changed. This
-        parameter is here for later compatibility with a future version of
-        the NFFT library which supports multiple precision, as available with
-        FFTW.
 
         **Precomputation flags**
 
@@ -487,65 +334,58 @@ cdef class NFFT:
 
     # where the C-related content of the class needs to be cleaned
     def __dealloc__(self):
-        if self.__plan != NULL:
-            self.__nfft_finalize(self.__plan)
-            free(self.__plan)
+        nfft_finalize(&self._plan)
 
     cpdef precompute(self):
         '''
         Precomputes the NFFT plan internals.
 
         .. warning::
-            The nodes :attr:`~pynfft.NFFT.x` must be initialized before
+            The nodes :attr:`pynfft.NFFT.x` must be initialized before
             precomputing.
         '''
-        if self.__plan != NULL:
-            with nogil:
-                self.__nfft_precompute(self.__plan)
+        with nogil:
+            nfft_precompute_one_psi(&self._plan)
 
     cpdef trafo(self):
         '''
         Performs the forward NFFT.
 
-        Reads :attr:`~pynfft.NFFT.f_hat` and stores the result in
-        :attr:`~pynfft.NFFT.f`.
+        Reads :attr:`pynfft.NFFT.f_hat` and stores the result in
+        :attr:`pynfft.NFFT.f`.
         '''
-        if self.__plan != NULL:
-            with nogil:
-                self.__nfft_trafo(self.__plan)
+        with nogil:
+            nfft_trafo(&self._plan)
 
     cpdef trafo_direct(self):
         '''
         Performs the forward NDFT.
 
-        Reads :attr:`~pynfft.NFFT.f_hat` and stores the result in
-        :attr:`~pynfft.NFFT.f`.
+        Reads :attr:`pynfft.NFFT.f_hat` and stores the result in
+        :attr:`pynfft.NFFT.f`.
         '''
-        if self.__plan != NULL:
-            with nogil:
-                self.__nfft_trafo_direct(self.__plan)
+        with nogil:
+             nfft_trafo_direct(&self._plan)
 
     cpdef adjoint(self):
         '''
         Performs the adjoint NFFT.
 
-        Reads :attr:`~pynfft.NFFT.f` and stores the result in
-        :attr:`~pynfft.NFFT.f_hat`.
+        Reads :attr:`pynfft.NFFT.f` and stores the result in
+        :attr:`pynfft.NFFT.f_hat`.
         '''
-        if self.__plan != NULL:
-            with nogil:
-                self.__nfft_adjoint(self.__plan)
+        with nogil:
+            nfft_adjoint(&self._plan)
 
     cpdef adjoint_direct(self):
         '''
         Performs the adjoint NDFT.
 
-        Reads :attr:`~pynfft.NFFT.f` and stores the result in
-        :attr:`~pynfft.NFFT.f_hat`.
+        Reads :attr:`pynfft.NFFT.f` and stores the result in
+        :attr:`pynfft.NFFT.f_hat`.
         '''
-        if self.__plan != NULL:
-            with nogil:
-                self.__nfft_adjoint_direct(self.__plan)
+        with nogil:
+             nfft_adjoint_direct(&self._plan)
 
     def __get_f(self):
         '''
@@ -627,6 +467,253 @@ cdef class NFFT:
         return self._n
 
     n = property(__get_n)
+
+    def __get_dtype(self):
+        '''
+        The floating precision.
+        '''
+        return self._dtype
+
+    dtype = property(__get_dtype)
+
+    def __get_flags(self):
+        '''
+        The precomputation flags.
+        '''
+        return self._flags
+
+    flags = property(__get_flags)
+
+
+##########
+# Solver #
+##########
+
+cdef object solver_flags_dict
+solver_flags_dict = {
+    'LANDWEBER':LANDWEBER,
+    'STEEPEST_DESCENT':STEEPEST_DESCENT,
+    'CGNR':CGNR,
+    'CGNE':CGNE,
+    'NORMS_FOR_LANDWEBER':NORMS_FOR_LANDWEBER,
+    'PRECOMPUTE_WEIGHT':PRECOMPUTE_WEIGHT,
+    'PRECOMPUTE_DAMP':PRECOMPUTE_DAMP,
+    }
+solver_flags = solver_flags_dict.copy()
+
+cdef class Solver:
+    '''
+    Solver is a class for computing the adjoint NFFT iteratively. Using the
+    solver should theoretically lead to more accurate results, even with just
+    one iteration, than using :meth:`pynfft.NFFT.adjoint` or
+    :meth:`pynfft.NFFT.adjoint_direct`.
+
+    The instantiation requires a NFFT object used internally for the multiple
+    forward and adjoint NFFT performed. The class uses conjugate-gradient as
+    the default solver but alternative solvers can be specified.
+
+    Because the stopping conidition of the iterative computation may change
+    from one application to another, the implementation only let you carry
+    one iteration at a time with a call to
+    :meth:`pynfft.Solver.loop_one_step`. Initialization of the solver
+    is done by calling the :meth:`pynfft.Solver.before_loop` method.
+
+    The class exposes the internals of the solver through call to their
+    respective properties. They should be treated as read-only values.
+    '''
+    cdef solver_plan_complex _plan
+    cdef NFFT _nfft_plan
+    cdef object _w
+    cdef object _w_hat
+    cdef object _y
+    cdef object _f_hat_iter
+    cdef object _r_iter
+    cdef object _dtype
+    cdef object _flags
+
+    def __cinit__(self, NFFT nfft_plan, flags=None):
+
+        # support only double / double complex NFFT
+        # TODO: if support for multiple floating precision lands in the
+        # NFFT library, adapt this section to dynamically figure the
+        # real and complex dtypes
+        dtype_real = np.dtype('float64')
+        dtype_complex = np.dtype('complex128')
+
+        # convert tuple of litteral precomputation flags to its expected
+        # C-compatible value. Each flag is a power of 2, which allows to compute
+        # this value using BITOR operations.
+        cdef unsigned int _flags = 0
+        flags_used = ()
+
+        # sanity checks on user specified flags if any,
+        # else use default ones:
+        if flags is not None:
+            try:
+                flags = tuple(flags)
+            except:
+                flags = (flags,)
+            finally:
+                flags_used += flags
+        else:
+            flags_used += ('CGNR',)
+
+        for each_flag in flags_used:
+            try:
+                _flags |= solver_flags_dict[each_flag]
+            except KeyError:
+                raise ValueError('Invalid flag: ' + '\'' +
+                        each_flag + '\' is not a valid flag.')
+
+        # initialize plan
+        try:
+            solver_init_advanced_complex(&self._plan,
+                <nfft_mv_plan_complex*>&(nfft_plan._plan), _flags)
+        except:
+            raise MemoryError
+
+        self._nfft_plan = nfft_plan
+
+        cdef np.npy_intp shape[1]
+        cdef int M_total = nfft_plan._M_total
+        cdef int N_total = nfft_plan._N_total
+
+        if 'PRECOMPUTE_WEIGHT' in flags_used:
+            shape[0] = M_total
+            self._w = np.PyArray_SimpleNewFromData(1, shape,
+                np.NPY_FLOAT64, <void *>(self._plan.w))
+            self._w[:] = 1  # make sure weights are initialized
+        else:
+            self._w = None
+
+        if 'PRECOMPUTE_DAMP' in flags_used:
+            shape[0] = N_total
+            self._w_hat = np.PyArray_SimpleNewFromData(1, shape,
+                np.NPY_FLOAT64, <void *>(self._plan.w_hat))
+            self._w_hat[:] = 1  # make sure weights are initialized
+        else:
+            self._w_hat = None
+
+        shape[0] = M_total
+        self._y = np.PyArray_SimpleNewFromData(1, shape,
+            np.NPY_COMPLEX128, <void *>(self._plan.y))
+
+        shape[0] = N_total
+        self._f_hat_iter = np.PyArray_SimpleNewFromData(1, shape,
+            np.NPY_COMPLEX128, <void *>(self._plan.f_hat_iter))
+        self._f_hat_iter[:] = 0  # default initial guess
+
+        shape[0] = M_total
+        self._r_iter = np.PyArray_SimpleNewFromData(1, shape,
+            np.NPY_COMPLEX128, <void *>(self._plan.r_iter))
+
+        self._dtype = dtype_real
+        self._flags = flags_used
+
+
+    def __init__(self, nfft_plan, flags=None):
+        '''
+        :param plan: instance of NFFT.
+        :type plan: :class:`pynfft.NFFT`
+        :param flags: list of instantiation flags, see below.
+        :type flags: tuple
+
+        **Instantiation flags**
+
+        +---------------------+-----------------------------------------------------------------------------+
+        | Flag                | Description                                                                 |
+        +=====================+=============================================================================+
+        | LANDWEBER           | Use Landweber (Richardson) iteration.                                       |
+        +---------------------+-----------------------------------------------------------------------------+
+        | STEEPEST_DESCENT    | Use steepest descent iteration.                                             |
+        +---------------------+-----------------------------------------------------------------------------+
+        | CGNR                | Use conjugate gradient (normal equation of the 1st kind).                   |
+        +---------------------+-----------------------------------------------------------------------------+
+        | CGNE                | Use conjugate gradient (normal equation of the 2nd kind).                   |
+        +---------------------+-----------------------------------------------------------------------------+
+        | NORMS_FOR_LANDWEBER | Use Landweber iteration to compute the residual norm.                       |
+        +---------------------+-----------------------------------------------------------------------------+
+        | PRECOMPUTE_WEIGHT   | Weight the samples, e.g. to cope with varying sampling density.             |
+        +---------------------+-----------------------------------------------------------------------------+
+        | PRECOMPUTE_DAMP     | Weight the Fourier coefficients, e.g. to favour fast decaying coefficients. |
+        +---------------------+-----------------------------------------------------------------------------+
+
+        Default value is ``flags = ('CGNR',)``.
+        '''
+        pass
+
+    def __dealloc__(self):
+        solver_finalize_complex(&self._plan)
+
+    cpdef before_loop(self):
+        '''
+        Initialize solver internals.
+        '''
+        with nogil:
+            solver_before_loop_complex(&self._plan)
+
+    cpdef loop_one_step(self):
+        '''
+        Perform one iteration.
+        '''
+        with nogil:
+            solver_loop_one_step_complex(&self._plan)
+
+    def __get_w(self):
+        '''
+        Weighting factors.
+        '''
+        return self._w
+
+    def __set_w(self, new_w):
+        if self._w is not None:
+            self._w.ravel()[:] = new_w.ravel()[:]
+
+    w = property(__get_w, __set_w)
+
+    def __get_w_hat(self):
+        '''
+        Damping factors.
+        '''
+        return self._w_hat
+
+    def __set_w_hat(self, new_w_hat):
+        if self._w_hat is not None:
+            self._w_hat.ravel()[:] = new_w_hat.ravel()[:]
+
+    w_hat = property(__get_w_hat, __set_w_hat)
+
+    def __get_y(self):
+        '''
+        Right hand side, samples.
+        '''
+        return self._y
+
+    def __set_y(self, new_y):
+        if self._y is not None:
+            self._y.ravel()[:] = new_y.ravel()[:]
+
+    y = property(__get_y, __set_y)
+
+    def __get_f_hat_iter(self):
+        '''
+        Iterative solution.
+        '''
+        return self._f_hat_iter
+
+    def __set_f_hat_iter(self, new_f_hat_iter):
+        if self._f_hat_iter is not None:
+            self._f_hat_iter.ravel()[:] = new_f_hat_iter.ravel()[:]
+
+    f_hat_iter = property(__get_f_hat_iter, __set_f_hat_iter)
+
+    def __get_r_iter(self):
+        '''
+        Residual vector.
+        '''
+        return self._r_iter
+
+    r_iter = property(__get_r_iter)
 
     def __get_dtype(self):
         '''