+++ changelog.rst
@@ -1,6 +1,31 @@
 Changelog
 =========
 
+Changes in version 1.0
+----------------------
+
+    - simplify code base
+
+    - pynfft.nfft and pynfft.solver moved to pynfft
+
+    - revert support for multiple floating precision 
+
+    - fix missing documentation for the solver
+
+
+Changes in version 0.6
+----------------------
+
+    - pynfft.nfft: enable openmp support
+
+Changes in version 0.5
+----------------------
+
+    - pynfft.nfft: rewrite NFFT class internals to support multiple floating 
+      point precision, coming in a future version of libnfft3
+
+    - Documenation: first draft of the tutorial section
+
 Changes in version 0.4.1
 ------------------------